backend/requirements.txt
Flask
python-dotenv
psycopg2-binary
flask_sqlalchemy
Flask-Cors
pytest